+++ GLMakie/src/screen.jl
@@ -8,8 +8,7 @@ function renderloop end
 """
 ## Renderloop
 
-* `renderloop = GLMakie.renderloop`: sets a function `renderloop(::GLMakie.Screen)` which starts a renderloop for the screen.
-
+* `renderloop = GLMakie.renderloop`: Sets a function `renderloop(::GLMakie.Screen)` which starts a renderloop for the screen.
 
 !!! warning
     The keyword arguments below are not effective if `renderloop` isn't set to `GLMakie.renderloop`, unless implemented in a custom renderloop function:
@@ -18,6 +17,7 @@ function renderloop end
 * `vsync = false`: Whether to enable vsync for the window.
 * `render_on_demand = true`: If `true`, the scene will only be rendered if something has changed in it.
 * `framerate = 30.0`: Sets the currently rendered frames per second.
+* `px_per_unit = automatic`: Sets the ratio between the number of rendered pixels and the `Makie` resolution. It defaults to the value of `scalefactor` but may be any positive real number.
 
 ## GLFW window attributes
 * `float = false`: Whether the window should float above other windows.
@@ -28,6 +28,7 @@ function renderloop end
 * `debugging = false`: If `true`, starts the GLFW.Window/OpenGL context with debug output.
 * `monitor::Union{Nothing, GLFW.Monitor} = nothing`: Sets the monitor on which the window should be opened. If set to `nothing`, GLFW will decide which monitor to use.
 * `visible = true`: Whether or not the window should be visible when first created.
+* `scalefactor = automatic`: Sets the window scaling factor, such as `2.0` on HiDPI/Retina displays. It is set automatically based on the display, but may be any positive real number.
 
 ## Postprocessor
 * `oit = false`: Whether to enable order independent transparency for the window.
@@ -43,6 +44,7 @@ mutable struct ScreenConfig
     vsync::Bool
     render_on_demand::Bool
     framerate::Float64
+    px_per_unit::Union{Nothing, Float32}
 
     # GLFW window attributes
     float::Bool
@@ -53,6 +55,7 @@ mutable struct ScreenConfig
     debugging::Bool
     monitor::Union{Nothing, GLFW.Monitor}
     visible::Bool
+    scalefactor::Union{Nothing, Float32}
 
     # Postprocessor
     oit::Bool
@@ -67,6 +70,7 @@ mutable struct ScreenConfig
             vsync::Bool,
             render_on_demand::Bool,
             framerate::Number,
+            px_per_unit::Union{Makie.Automatic, Number},
             # GLFW window attributes
             float::Bool,
             focus_on_show::Bool,
@@ -76,6 +80,7 @@ mutable struct ScreenConfig
             debugging::Bool,
             monitor::Union{Nothing, GLFW.Monitor},
             visible::Bool,
+            scalefactor::Union{Makie.Automatic, Number},
 
             # Preprocessor
             oit::Bool,
@@ -89,6 +94,7 @@ mutable struct ScreenConfig
             vsync,
             render_on_demand,
             framerate,
+            px_per_unit isa Makie.Automatic ? nothing : Float32(px_per_unit),
             # GLFW window attributes
             float,
             focus_on_show,
@@ -98,6 +104,8 @@ mutable struct ScreenConfig
             debugging,
             monitor,
             visible,
+            scalefactor isa Makie.Automatic ? nothing : Float32(scalefactor),
+            # Preproccessor
             # Preprocessor
             oit,
             fxaa,
@@ -147,6 +155,7 @@ mutable struct Screen{GLWindow} <: MakieScreen
     config::Union{Nothing, ScreenConfig}
     stop_renderloop::Bool
     rendertask::Union{Task, Nothing}
+    px_per_unit::Observable{Float32}
 
     screen2scene::Dict{WeakRef, ScreenID}
     screens::Vector{ScreenArea}
@@ -157,6 +166,7 @@ mutable struct Screen{GLWindow} <: MakieScreen
     framecache::Matrix{RGB{N0f8}}
     render_tick::Observable{Nothing}
     window_open::Observable{Bool}
+    scalefactor::Observable{Float32}
 
     root_scene::Union{Scene, Nothing}
     reuse::Bool
@@ -185,10 +195,10 @@ mutable struct Screen{GLWindow} <: MakieScreen
         screen = new{GLWindow}(
             glscreen, shader_cache, framebuffer,
             config, stop_renderloop, rendertask,
-            screen2scene,
+            Observable(0f0), screen2scene,
             screens, renderlist, postprocessors, cache, cache2plot,
             Matrix{RGB{N0f8}}(undef, s), Observable(nothing),
-            Observable(true), nothing, reuse, true, false
+            Observable(true), Observable(0f0), nothing, reuse, true, false
         )
         push!(ALL_SCREENS, screen) # track all created screens
         return screen
@@ -213,6 +223,8 @@ function empty_screen(debugging::Bool; reuse=true)
 
         (GLFW.STENCIL_BITS, 0),
         (GLFW.AUX_BUFFERS,  0),
+
+        (GLFW.SCALE_TO_MONITOR, true),
     ]
     resolution = (10, 10)
     window = try
@@ -261,7 +273,9 @@ function empty_screen(debugging::Bool; reuse=true)
         Dict{UInt32, AbstractPlot}(),
         reuse,
     )
-    GLFW.SetWindowRefreshCallback(window, window -> refreshwindowcb(window, screen))
+    GLFW.SetWindowRefreshCallback(window, refreshwindowcb(screen))
+    GLFW.SetWindowContentScaleCallback(window, scalechangecb(screen))
+
     return screen
 end
 
@@ -276,6 +290,7 @@ function reopen!(screen::Screen)
     end
     @assert isempty(screen.window_open.listeners)
     screen.window_open[] = true
+    on(scalechangeobs(screen), screen.scalefactor)
     @assert isopen(screen)
     return screen
 end
@@ -307,8 +322,6 @@ function singleton_screen(debugging::Bool)
     return reopen!(screen)
 end
 
-const GLFW_FOCUS_ON_SHOW = 0x0002000C
-
 function Makie.apply_screen_config!(screen::Screen, config::ScreenConfig, scene::Scene, args...)
     apply_config!(screen, config)
 end
@@ -317,7 +330,7 @@ function apply_config!(screen::Screen, config::ScreenConfig; start_renderloop::B
     @debug("Applying screen config! to existing screen")
     glw = screen.glscreen
     ShaderAbstractions.switch_context!(glw)
-    GLFW.SetWindowAttrib(glw, GLFW_FOCUS_ON_SHOW, config.focus_on_show)
+    GLFW.SetWindowAttrib(glw, GLFW.FOCUS_ON_SHOW, config.focus_on_show)
     GLFW.SetWindowAttrib(glw, GLFW.DECORATED, config.decorated)
     GLFW.SetWindowAttrib(glw, GLFW.FLOATING, config.float)
     GLFW.SetWindowTitle(glw, config.title)
@@ -325,6 +338,8 @@ function apply_config!(screen::Screen, config::ScreenConfig; start_renderloop::B
     if !isnothing(config.monitor)
         GLFW.SetWindowMonitor(glw, config.monitor)
     end
+    screen.scalefactor[] = !isnothing(config.scalefactor) ? config.scalefactor : scale_factor(glw)
+    screen.px_per_unit[] = !isnothing(config.px_per_unit) ? config.px_per_unit : screen.scalefactor[]
 
     function replace_processor!(postprocessor, idx)
         fb = screen.framebuffer
@@ -361,10 +376,10 @@ function Screen(;
     # Screen config is managed by the current active theme, so managed by Makie
     config = Makie.merge_screen_config(ScreenConfig, screen_config)
     screen = screen_from_pool(config.debugging)
+    apply_config!(screen, config; start_renderloop=start_renderloop)
     if !isnothing(resolution)
         resize!(screen, resolution...)
     end
-    apply_config!(screen, config; start_renderloop=start_renderloop)
     return screen
 end
 
@@ -563,7 +578,10 @@ function destroy!(screen::Screen)
     # otherwise, during rendertask clean up we may run into a destroyed window
     wait(screen)
     screen.rendertask = nothing
-    destroy!(screen.glscreen)
+    window = screen.glscreen
+    GLFW.SetWindowRefreshCallback(window, nothing)
+    GLFW.SetWindowContentScaleCallback(window, nothing)
+    destroy!(window)
     # Since those are sets, we can just delete them from there, even if they weren't in there (e.g. reuse=false)
     delete!(SCREEN_REUSE_POOL, screen)
     delete!(ALL_SCREENS, screen)
@@ -587,6 +605,8 @@ function Base.close(screen::Screen; reuse=true)
         screen.window_open[] = false
     end
     empty!(screen)
+    Observables.clear(screen.px_per_unit)
+    Observables.clear(screen.scalefactor)
     if reuse && screen.reuse
         @debug("reusing screen!")
         push!(SCREEN_REUSE_POOL, screen)
@@ -614,24 +634,30 @@ function closeall()
     return
 end
 
-function resize_native!(window::GLFW.Window, resolution...)
-    if isopen(window)
-        ShaderAbstractions.switch_context!(window)
-        oldsize = windowsize(window)
-        retina_scale = retina_scaling_factor(window)
-        w, h = resolution ./ retina_scale
-        if oldsize == (w, h)
-            return
-        end
-        GLFW.SetWindowSize(window, round(Int, w), round(Int, h))
-    end
-end
+function Base.resize!(screen::Screen, w::Int, h::Int)
+    window = to_native(screen)
+    (w > 0 && h > 0 && isopen(window)) || return nothing
 
-function Base.resize!(screen::Screen, w, h)
-    nw = to_native(screen)
-    resize_native!(nw, w, h)
-    fb = screen.framebuffer
-    resize!(fb, (w, h))
+    # Resize the window which appears on the user desktop (if necessary).
+    #
+    # On OSX with a Retina display, the window size is given in logical dimensions and
+    # is automatically scaled by the OS. To support arbitrary scale factors, we must account
+    # for the native scale factor when calculating the effective scaling to apply.
+    #
+    # On Linux and Windows, scale from the logical size to the pixel size.
+    ShaderAbstractions.switch_context!(window)
+    winscale = screen.scalefactor[] / (@static Sys.isapple() ? scale_factor(window) : 1)
+    winw, winh = round.(Int, winscale .* (w, h))
+    if window_size(window) != (winw, winh)
+        GLFW.SetWindowSize(window, winw, winh)
+    end
+
+    # Then resize the underlying rendering framebuffers as well, which can be scaled
+    # independently of the window scale factor.
+    fbscale = screen.px_per_unit[]
+    fbw, fbh = round.(Int, fbscale .* (w, h))
+    resize!(screen.framebuffer, fbw, fbh)
+    return nothing
 end
 
 function fast_color_data!(dest::Array{RGB{N0f8}, 2}, source::Texture{T, 2}) where T
@@ -679,8 +705,7 @@ function Makie.colorbuffer(screen::Screen, format::Makie.ImageStorageFormat = Ma
     # polling may change window size, when its bigger than monitor!
     # we still need to poll though, to get all the newest events!
     # GLFW.PollEvents()
-    # keep current buffer size to allows larger-than-window renders
-    render_frame(screen, resize_buffers=false) # let it render
+    render_frame(screen, resize_buffers=true) # let it render
     glFinish() # block until opengl is done rendering
     if size(ctex) != size(screen.framecache)
         screen.framecache = Matrix{RGB{N0f8}}(undef, size(ctex))
@@ -804,12 +829,32 @@ function set_framerate!(screen::Screen, fps=30)
     screen.config.framerate = fps
 end
 
-function refreshwindowcb(window, screen)
+function refreshwindowcb(screen, window)
     screen.render_tick[] = nothing
     render_frame(screen)
     GLFW.SwapBuffers(window)
     return
 end
+refreshwindowcb(screen) = window -> refreshwindowcb(screen, window)
+
+function scalechangecb(screen, window, xscale, yscale)
+    sf = min(xscale, yscale)
+    if isnothing(screen.config.px_per_unit) && screen.scalefactor[] == screen.px_per_unit[]
+        screen.px_per_unit[] = sf
+    end
+    screen.scalefactor[] = sf
+    return
+end
+scalechangecb(screen) = (window, xscale, yscale) -> scalechangecb(screen, window, xscale, yscale)
+
+function scalechangeobs(screen, _)
+    if !isnothing(screen.root_scene)
+        resize!(screen, size(screen.root_scene)...)
+    end
+    return nothing
+end
+scalechangeobs(screen) = scalefactor -> scalechangeobs(screen, scalefactor)
+
 
 # TODO add render_tick event to scene events
 function vsynced_renderloop(screen)