+++ src/theming.jl
@@ -85,6 +85,7 @@ const minimal_default = Attributes(
         vsync = false,
         render_on_demand = true,
         framerate = 30.0,
+        px_per_unit = automatic,
 
         # GLFW window attributes
         float = false,
@@ -95,6 +96,7 @@ const minimal_default = Attributes(
         debugging = false,
         monitor = nothing,
         visible = true,
+        scalefactor = automatic,
 
         # Postproccessor
         oit = true,